issue #2 resolved
updated package names
