+++ pom.xml
@@ -10,9 +10,9 @@
     <relativePath /> <!-- lookup parent from repository -->
   </parent>
   <groupId>com.coderscampus</groupId>
-  <artifactId>SpringSecurityJWTDemo</artifactId>
+  <artifactId>ThePorscheProject</artifactId>
   <version>0.0.1-SNAPSHOT</version>
-  <name>SpringSecurityJWTDemo</name>
+  <name>ThePorscheProject</name>
   <description>Demo project for Spring Boot</description>
   <properties>
     <java.version>17</java.version>
