+++ src/main/java/com/ThePorscheProject/dao/request/RefreshTokenRequest.java
@@ -1,0 +1,5 @@
+package com.ThePorscheProject.dao.request;
+
+public record RefreshTokenRequest(String refreshToken) {
+
+}
