+++ src/main/java/com/ThePorscheProject/dao/request/SignInRequest.java
@@ -1,0 +1,5 @@
+package com.ThePorscheProject.dao.request;
+
+public record SignInRequest(String email, String password) {
+
+}
