+++ src/main/java/com/ThePorscheProject/dao/request/SignUpRequest.java
@@ -1,0 +1,11 @@
+package com.ThePorscheProject.dao.request;
+
+import java.util.Optional;
+
+public record SignUpRequest(String email,
+                            String password,
+                            String firstName,
+                            String lastName,
+                            Optional<String> authorityOpt) {
+
+}
