+++ src/main/java/com/ThePorscheProject/dao/response/JwtAuthenticationResponse.java
@@ -1,0 +1,5 @@
+package com.ThePorscheProject.dao.response;
+
+public record JwtAuthenticationResponse(String token, String refreshToken) {
+
+}
