+++ src/main/java/com/ThePorscheProject/dao/response/TokenRefreshResponse.java
@@ -1,0 +1,5 @@
+package com.ThePorscheProject.dao.response;
+
+public record TokenRefreshResponse(String accessToken, String refreshToken) {
+
+}
