+++ src/main/java/com/ThePorscheProject/domain/Product.java
@@ -1,0 +1,10 @@
+package com.ThePorscheProject.domain;
+
+import java.math.BigDecimal;
+
+public record Product (Integer id, String name, BigDecimal price){
+    
+	/*
+	 * THIS CLASS IS FOR TESTING PURPOSES ONLY
+	 */
+}
