+++ src/main/java/com/ThePorscheProject/domain/Role.java
@@ -1,0 +1,6 @@
+package com.ThePorscheProject.domain;
+
+public enum Role {
+    USER,
+    ADMIN
+}
