+++ src/main/java/com/ThePorscheProject/domain/User.java
@@ -1,0 +1,142 @@
+package com.ThePorscheProject.domain;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import jakarta.persistence.*;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+@Entity
+@Table(name = "users")
+
+public class User implements UserDetails {
+    private static final long serialVersionUID = 2025389852147750927L;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    private String firstName;
+    private String lastName;
+    private String email;
+    private String password;
+    @OneToMany(mappedBy = "user", fetch = FetchType.EAGER, cascade = CascadeType.ALL)
+    @JsonIgnoreProperties("user")
+    private List<Authority> authorities = new ArrayList<>();
+    
+    /*
+     * enable if you want to have the confirm password to be checked in the back end + their corresponding getters and setters
+     * private String confirmPassword;
+     */
+    
+    
+    @Override
+    public Collection<Authority> getAuthorities() {
+        return authorities;
+    }
+
+    public void setAuthorities(List<Authority> authorities) {
+        this.authorities = authorities;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    @Override
+    public String getUsername() {
+        // email in our case
+        return email;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+
+    @Override
+    public String getPassword() {
+        return this.password;
+    }
+    
+    public User firstName(String firstName) {
+        this.firstName = firstName;
+        return this;
+    }
+    
+    public User lastName(String lastName) {
+        this.lastName = lastName;
+        return this;
+    }
+    
+    public User email(String email) {
+        this.email = email;
+        return this;
+    }
+
+    public User authority(String authority) {
+        Authority auth = new Authority(authority, this);
+        this.getAuthorities().add(auth);
+        return this;
+    }
+    
+    public User password(String password) {
+        this.password = password;
+        return this;
+    }
+    
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public Integer getId() {
+        return id;
+    }
+
+
+    public User build () {
+        return this;
+    }
+
+}
