+++ src/main/java/com/ThePorscheProject/repository/RefreshTokenRepository.java
@@ -1,0 +1,18 @@
+package com.ThePorscheProject.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.stereotype.Repository;
+
+import com.ThePorscheProject.domain.RefreshToken;
+import com.ThePorscheProject.domain.User;
+
+@Repository
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+    Optional<RefreshToken> findByToken(String token);
+
+    @Modifying
+    int deleteByUser(User user);
+}
