+++ src/main/java/com/ThePorscheProject/repository/UserRepository.java
@@ -1,0 +1,15 @@
+package com.ThePorscheProject.repository;
+
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.ThePorscheProject.domain.User;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Integer> {
+    // Since email is unique, we'll find users by email
+    Optional<User> findByEmail(String email);
+}
