+++ src/main/java/com/ThePorscheProject/security/AuthenticationService.java
@@ -1,0 +1,10 @@
+package com.ThePorscheProject.security;
+
+import com.ThePorscheProject.dao.request.SignInRequest;
+import com.ThePorscheProject.dao.request.SignUpRequest;
+import com.ThePorscheProject.dao.response.JwtAuthenticationResponse;
+
+public interface AuthenticationService {
+    JwtAuthenticationResponse signup(SignUpRequest request);
+    JwtAuthenticationResponse signin(SignInRequest request);
+}
