+++ src/main/java/com/ThePorscheProject/security/JwtService.java
@@ -1,0 +1,11 @@
+package com.ThePorscheProject.security;
+
+import org.springframework.security.core.userdetails.UserDetails;
+
+public interface JwtService {
+    String extractUserName(String token);
+
+    String generateToken(UserDetails userDetails);
+
+    boolean isTokenValid(String token, UserDetails userDetails);
+}
