+++ src/main/java/com/ThePorscheProject/service/UserService.java
@@ -1,0 +1,12 @@
+package com.ThePorscheProject.service;
+
+import org.springframework.security.core.userdetails.UserDetailsService;
+
+import com.ThePorscheProject.domain.User;
+
+import java.util.List;
+
+public interface UserService {
+    UserDetailsService userDetailsService();
+    List<User> findAll();
+}
