+++ src/main/java/com/ThePorscheProject/util/CookieUtils.java
@@ -1,0 +1,18 @@
+package com.ThePorscheProject.util;
+
+import jakarta.servlet.http.Cookie;
+
+public class CookieUtils {
+
+	public static Cookie createAccessTokenCookie(String value) {
+		Cookie accessTokenCookie = new Cookie("accessToken", value);
+    	
+		return accessTokenCookie;
+	}
+	
+	public static Cookie createRefreshTokenCookie(String value) {
+		Cookie refreshTokenCookie = new Cookie("refreshToken", value);
+		
+		return refreshTokenCookie;
+	}		
+}
