--- src/main/java/com/coderscampus/SpringSecurityJWTDemo/SpringSecurityJwtDemoApplication.java
@@ -1,14 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringSecurityJwtDemoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringSecurityJwtDemoApplication.class, args);
-	}
-	
-	//test
-}
