--- src/main/java/com/coderscampus/SpringSecurityJWTDemo/dao/request/RefreshTokenRequest.java
@@ -1,5 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo.dao.request;
-
-public record RefreshTokenRequest(String refreshToken) {
-
-}
