--- src/main/java/com/coderscampus/SpringSecurityJWTDemo/dao/request/SignInRequest.java
@@ -1,5 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo.dao.request;
-
-public record SignInRequest(String email, String password) {
-
-}
