--- src/main/java/com/coderscampus/SpringSecurityJWTDemo/dao/request/SignUpRequest.java
@@ -1,11 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo.dao.request;
-
-import java.util.Optional;
-
-public record SignUpRequest(String email,
-                            String password,
-                            String firstName,
-                            String lastName,
-                            Optional<String> authorityOpt) {
-
-}
