--- src/main/java/com/coderscampus/SpringSecurityJWTDemo/dao/response/JwtAuthenticationResponse.java
@@ -1,5 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo.dao.response;
-
-public record JwtAuthenticationResponse(String token, String refreshToken) {
-
-}
