--- src/main/java/com/coderscampus/SpringSecurityJWTDemo/dao/response/TokenRefreshResponse.java
@@ -1,5 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo.dao.response;
-
-public record TokenRefreshResponse(String accessToken, String refreshToken) {
-
-}
