--- src/main/java/com/coderscampus/SpringSecurityJWTDemo/domain/Product.java
@@ -1,10 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo.domain;
-
-import java.math.BigDecimal;
-
-public record Product (Integer id, String name, BigDecimal price){
-    
-	/*
-	 * THIS CLASS IS FOR TESTING PURPOSES ONLY
-	 */
-}
