--- src/main/java/com/coderscampus/SpringSecurityJWTDemo/domain/Role.java
@@ -1,6 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo.domain;
-
-public enum Role {
-    USER,
-    ADMIN
-}
