--- src/main/java/com/coderscampus/SpringSecurityJWTDemo/repository/RefreshTokenRepository.java
@@ -1,18 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.stereotype.Repository;
-
-import com.coderscampus.SpringSecurityJWTDemo.domain.RefreshToken;
-import com.coderscampus.SpringSecurityJWTDemo.domain.User;
-
-@Repository
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-    Optional<RefreshToken> findByToken(String token);
-
-    @Modifying
-    int deleteByUser(User user);
-}
