--- src/main/java/com/coderscampus/SpringSecurityJWTDemo/repository/UserRepository.java
@@ -1,15 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo.repository;
-
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.coderscampus.SpringSecurityJWTDemo.domain.User;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-    // Since email is unique, we'll find users by email
-    Optional<User> findByEmail(String email);
-}
