--- src/main/java/com/coderscampus/SpringSecurityJWTDemo/security/JwtService.java
@@ -1,11 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo.security;
-
-import org.springframework.security.core.userdetails.UserDetails;
-
-public interface JwtService {
-    String extractUserName(String token);
-
-    String generateToken(UserDetails userDetails);
-
-    boolean isTokenValid(String token, UserDetails userDetails);
-}
