--- src/main/java/com/coderscampus/SpringSecurityJWTDemo/service/UserService.java
@@ -1,11 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo.service;
-
-import com.coderscampus.SpringSecurityJWTDemo.domain.User;
-import org.springframework.security.core.userdetails.UserDetailsService;
-
-import java.util.List;
-
-public interface UserService {
-    UserDetailsService userDetailsService();
-    List<User> findAll();
-}
