--- src/main/java/com/coderscampus/SpringSecurityJWTDemo/util/CookieUtils.java
@@ -1,18 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo.util;
-
-import jakarta.servlet.http.Cookie;
-
-public class CookieUtils {
-
-	public static Cookie createAccessTokenCookie(String value) {
-		Cookie accessTokenCookie = new Cookie("accessToken", value);
-    	
-		return accessTokenCookie;
-	}
-	
-	public static Cookie createRefreshTokenCookie(String value) {
-		Cookie refreshTokenCookie = new Cookie("refreshToken", value);
-		
-		return refreshTokenCookie;
-	}		
-}
