+++ src/test/java/com/ThePorscheProject/SpringSecurityJwtDemoApplicationTests.java
@@ -1,0 +1,13 @@
+package com.ThePorscheProject;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ThePorscheProjectApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
