--- src/test/java/com/coderscampus/SpringSecurityJWTDemo/SpringSecurityJwtDemoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.coderscampus.SpringSecurityJWTDemo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringSecurityJwtDemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
